refactor(studio): extract usePreviewInteraction from useDomEditSession
